+++ category/Foundation/NSArray+RFKit.h
@@ -12,8 +12,6 @@
 #import <Foundation/Foundation.h>
 
 @interface NSArray (RFKit)
-
-- (id)firstObject;
 @end
 
 @interface NSMutableArray (RFKit)