+++ category/Foundation/NSArray+RFKit.m
@@ -4,10 +4,6 @@
 
 @implementation NSArray (RFKit)
 
-- (id)firstObject {
-    return (self!=nil && self.count>0) ? [self objectAtIndex:0] : nil;
-}
-
 @end
 
 @implementation NSMutableArray (RFKit)