+++ category/UIKit/UIViewController+RFKit.m
@@ -4,13 +4,6 @@
 
 @implementation UIViewController (RFKit)
 
-- (void)setNavTitle:(NSString *)title back:(NSString *)backTitle{
-	self.title = title;
-	UIBarButtonItem * tmpBack = [[UIBarButtonItem alloc] initWithTitle:backTitle style:UIBarButtonItemStylePlain target:self action:nil];
-	self.navigationItem.backBarButtonItem = tmpBack;
-	RF_RELEASE_OBJ(tmpBack);
-}
-
 + (UIViewController *)rootViewControllerWhichCanPresentModalViewController {
     UIViewController *vc = ([UIApplication sharedApplication].keyWindow.rootViewController)? : [(UIWindow *)[[UIApplication sharedApplication].windows firstObject] rootViewController];